feat: APP-15 packageId for animated views updated
