+++ designsystem/animatedsplashview/build.gradle.kts
@@ -1,10 +1,10 @@
 plugins {
     id("com.android.library")
-    id("org.jetbrains.kotlin.android")
+    kotlin("android")
 }
 
 android {
-    namespace = "com.github.mobdev778.yusupova.animatedsplashview"
+    namespace = "com.github.mobdev778.yusupova.designsystem.animatedsplashview"
     compileSdk = ProjectVersions.compileSdk
 
     defaultConfig {
