+++ designsystem/animatedsplashview/src/main/java/com/github/mobdev778/yusupova/designsystem/animatedsplashview/figure/XYMap.kt
@@ -1,4 +1,4 @@
-package com.github.mobdev778.yusupova.animatedtextview.figure
+package com.github.mobdev778.yusupova.designsystem.animatedsplashview.figure
 
 internal class XYMap<T> {
 
@@ -22,8 +22,26 @@ internal class XYMap<T> {
         return xyValueMap.size
     }
 
+    fun keyIterator(): Iterator<Pair<Int, Int>> {
+        return XYMapIterator(xyValueMap.keys.iterator())
+    }
+
     fun valueIterator(): Iterator<T> {
         return xyValueMap.values.iterator()
+    }
+
+    private class XYMapIterator(private val iterator: Iterator<Int>): Iterator<Pair<Int, Int>> {
+
+        override fun hasNext(): Boolean {
+            return iterator.hasNext()
+        }
+
+        override fun next(): Pair<Int, Int> {
+            val next = iterator.next()
+            val x = next.shr(X_BIT_SHIFT).and(0xFFF)
+            val y = next.and(0xFFF)
+            return Pair(x, y)
+        }
     }
 
     companion object {
