--- designsystem/animatedtextview/src/main/java/com/github/mobdev778/yusupova/animatedtextview/figure/Marker.kt
@@ -1,3 +1,0 @@
-package com.github.mobdev778.yusupova.animatedtextview.figure
-
-internal data class Marker(var number: Int)
