+++ designsystem/animatedtextview/src/main/java/com/github/mobdev778/yusupova/designsystem/animatedtextview/AnimatedTextMode.kt
@@ -1,4 +1,4 @@
-package com.github.mobdev778.yusupova.animatedtextview
+package com.github.mobdev778.yusupova.designsystem.animatedtextview
 
 enum class AnimatedTextMode {
     DEFAULT,
