+++ designsystem/animatedtextview/src/main/java/com/github/mobdev778/yusupova/designsystem/animatedtextview/AnimatedTextView.kt
@@ -1,4 +1,4 @@
-package com.github.mobdev778.yusupova.animatedtextview
+package com.github.mobdev778.yusupova.designsystem.animatedtextview
 
 import android.graphics.Typeface
 import android.text.Layout
@@ -30,7 +30,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.asCoroutineDispatcher
 import kotlinx.coroutines.launch
