+++ designsystem/animatedtextview/src/main/java/com/github/mobdev778/yusupova/designsystem/animatedtextview/figure/ImmutableFigure.kt
@@ -1,4 +1,4 @@
-package com.github.mobdev778.yusupova.animatedtextview.figure
+package com.github.mobdev778.yusupova.designsystem.animatedtextview.figure
 
 internal class ImmutableFigure(
     private val points: XYMap<Point>,
