+++ designsystem/animatedtextview/src/main/java/com/github/mobdev778/yusupova/designsystem/animatedtextview/figure/Marker.kt
@@ -1,0 +1,3 @@
+package com.github.mobdev778.yusupova.designsystem.animatedtextview.figure
+
+internal data class Marker(var number: Int)
